+++ client/tws_foundation_client/lib/src/services/security/account_service.dart
@@ -16,4 +16,16 @@ final class AccountService extends AccountsServiceBase {
     CSMActEffect actEffect = await twsPost('view', options, auth: auth);
     return MainResolver<SetViewOut<Account>>(actEffect);
   }
+
+  @override
+  Effect<SetBatchOut<Account>> create(List<Account> accounts, String auth) async {
+    CSMActEffect actEffect = await twsPostList('create', accounts, auth: auth);
+    return MainResolver<SetBatchOut<Account>>(actEffect);
+  }
+
+  @override
+  Effect<RecordUpdateOut<Account>> update(Account account, String auth) async {
+    CSMActEffect actEffect = await twsPost('update', account, auth: auth);
+    return MainResolver<RecordUpdateOut<Account>>(actEffect);
+  }
 }